move ct errors out of cttc
